+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
 
-## [Initial Version] - 2022-10-21
+## [Initial Version] - 2022-12-23
+
+- Create card
+- Delete card
+- Flip card
+- List and search cards